This is a test for dbase_add_record().
It opens the dbase file, adds records, closes it,
then RE-opens it and reads it all in using dbase_get_record_with_names(),
var_dump()'s the result, then closes and unlinks the dbf I created.

PROBLEM: Right now, it uses "/tmp" to create the .dbf, so it won't work on
Win32 (AFAIK).