Fixed bug #75326 (https://qa.php.net/tested-methods.php is outdated)
This is done by adding a new function (get_active_branches()), that returns a constant array with active versions (hence the added note in release-qa.php).

The write-test.php page now has a new paragraph listing active branches and linking to GCOV instead of the old and outdated coverage data we had stored on QA (which almost managed to celebrate a 10 year birthday).