Fix recognition of test filenames
The filenames of submitted test runs are determined from each line of
the test results, which has the format `testname [filename]`.  If the
testname contains a bracketed phrase, the filename is not correctly
recognized.  We fix this by anchoring the pattern to the end of the
line (but allow for trailing whitespace).

This is supposed to fix the resulting mess that currently can be seen
on <http://qa.php.net/reports/run_tests.php?version=7.2.0> for new bug
reports.