Fix recognition of test filenames
The filenames of submitted test runs are determined from each line of
the test results, which has the format `testname [filename]` for failed
tests.  If the testname contains a bracketed phrase, the filename is
not correctly recognized.  Expected failing tests have the format
`testname [filename] XFAIL REASON: reason`, though.  We fix this by
asserting that the bracketed filename is either at the end of the
string or that it's been followed by `XFAIL`.

We also cater to output of redirected tests, which add lines with the
format `via [filename]`, which we have to ignore.

This is supposed to fix the resulting mess that currently can be seen
on <http://qa.php.net/reports/run_tests.php?version=7.2.0> for new bug
reports.

However, this kind of filename recognition is still fragile, especially
with regard to potential changes of the run-tests.php output.  It
might be sensible to change it so it can more reliably be parsed.  We'd
still would have to cater to tests results produced with the present
test runner, though.