Refactor list_builds.php
I also removed the table with the "Latest:" entry as the following
table has the latest entry as it's first item anyway.

Note also that the prior implementation had an edge case where if
two revision directories had the same mtime, only one would be listed.
I resolved that by keying the array on the revision ID,
and using the mtime as the value for rsort()ing.