Remove server-tests.php related documentation
Since server-tests.php has been removed[1], we remove the related
documentation as well.  While the documentation might still be somewhat
useful for maintaining tests in PHP-7.3 and externally, we feel that
the benefits of not confusing test writers with obsolete information
outweigh the disadvantage of possibly having to look up the info
elsewhere.

Especially, we can fix the ENV section documentation for run-tests.php,
which assumes another format than server-tests.php.

[1] <http://git.php.net/?p=php-src.git;a=commit;h=4f36acb9e65935aa657f1f22e2320a401bdbdad3>